Remove interface
